Issue #18931: Move dialog adapter to collections package
