+++ app/src/main/java/org/mozilla/fenix/tabtray/TabTrayDialogFragment.kt
@@ -50,6 +50,7 @@ import mozilla.components.support.utils.ext.right
 import org.mozilla.fenix.HomeActivity
 import org.mozilla.fenix.NavGraphDirections
 import org.mozilla.fenix.R
+import org.mozilla.fenix.collections.CollectionsListAdapter
 import org.mozilla.fenix.components.FenixSnackbar
 import org.mozilla.fenix.components.StoreProvider
 import org.mozilla.fenix.components.TabCollectionStorage
