+++ app/src/test/java/org/mozilla/fenix/collections/CollectionsListAdapterTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.tabtray
+package org.mozilla.fenix.collections
 
 import android.widget.FrameLayout
 import io.mockk.mockk
